Drone phasing bug sixed
A small bug where the drone would "phase" through geometry has been fixed.
The bug would occuer if the drone was traveling at certain speeds or reching a certain momentum.
To Stop this a raycast function has been implemented to check if the drone is near any collision surface and resets the speed and momentum to zero.

Also made small changes in the URP settings.