Treasure rendering preview
Made a small system to instantiate treasure models when a Scriptable Object (SO) has been chosen.
This means that the gameobjects will be dynamicly created in the game world in edit mode and changed if the SO changes.

Also made some changes and edits in the existing treasures and made a new one.