TreasureSystem re-rework
The treasure system is now using generics instead for decoupling some duplicated logic.
There is also two new ScriptableObjects (SO), one acting as a base holding variables the other SO's are going to need. While the second one is dedicated for treasure chests.
This also includes a base Treasure class , so the radar can still detect the objects without haing to specify a specifik generic.